+++ src/components/Constructors/SingleConstructor.tsx
@@ -33,7 +33,7 @@ export default function SingleConstructor(props: {name: String, chief: String, p
             <div className="px-6 pt-4 pb-2 bg-[#161616]">
               <span className="inline-block bg-gray-200 rounded-full px-3 py-1 text-sm font-semibold text-gray-700 mr-2 mb-2">{props.chasis}</span>
               <span className="inline-block bg-gray-200 rounded-full px-3 py-1 text-sm font-semibold text-gray-700 mr-2 mb-2">{props.first}</span>
-              <span className="inline-block bg-gray-200 rounded-full px-3 py-1 text-sm font-semibold text-gray-700 mr-2 mb-2">{props.world} Points</span>
+              <span className="inline-block bg-gray-200 rounded-full px-3 py-1 text-sm font-semibold text-gray-700 mr-2 mb-2">{props.world} Championships</span>
             </div>
           </div>
         </div>
